+++ providers/capv/upgrade/capv_test.go
@@ -3,8 +3,8 @@ package upgrade
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v2/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v2/internal/upgrade"
+	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v3/internal/upgrade"
 )
 
 var _ = Describe("Basic upgrade test", Ordered, func() {