+++ providers/capvcd/upgrade/capvcd_test.go
@@ -5,10 +5,10 @@ import (
 
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v2/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v2/internal/state"
-	"github.com/giantswarm/cluster-test-suites/v2/internal/timeout"
-	"github.com/giantswarm/cluster-test-suites/v2/internal/upgrade"
+	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v3/internal/state"
+	"github.com/giantswarm/cluster-test-suites/v3/internal/timeout"
+	"github.com/giantswarm/cluster-test-suites/v3/internal/upgrade"
 )
 
 var _ = Describe("Basic upgrade test", Ordered, func() {