+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Dependencies
 
-- Update `clustertest` to v3.0.0 (pending release) with breaking changes removing default-apps support.
+- Update `clustertest` to v3.0.0 with breaking changes removing default-apps support.
+- Update `cluster-standup-teardown` to v3.1.0 (pending release) with removal of default-apps values files and interface changes.
 
 ## [2.7.1] - 2025-12-18